Convert the tutorial project to the new project structure
