--- src/fuddly/projects/generic/__init__.py
@@ -1,1 +1,0 @@
-
