--- src/fuddly/projects/specific/__init__.py
@@ -1,1 +1,0 @@
-
