+++ src/fuddly/projects/tuto/__init__.py
@@ -1,0 +1,2 @@
+from .prj import (project, logger)
+from .targets import targets
