+++ src/fuddly/projects/tuto/scripts/README
@@ -1,0 +1,3 @@
+In this directory you can add scripts used to run your projects.
+
+See the fuddly documentaion for more information on the subject
